More careful writing

+++ note-collapse.rb
@@ -23,8 +23,10 @@ def parsefile(doc)
 end
 
 puts "#{startNodes.length} posts"
-puts "time | left  | remaining "
+puts "time | left  | remain | +space"
 count = 0
+space_in = 0
+space_out = 0
 
 startNodes.each { | x |
   count += 1
@@ -32,10 +34,13 @@ startNodes.each { | x |
   if count % 10 == 0
     duration = Time.new - $start
     ttl = (duration / (count.to_f / startNodes.length.to_f)).to_i
-    togo = "%d:%02d" % [(ttl / 60).to_i, ttl % 60]
+    togo = "%02d:%02d" % [(ttl / 60).to_i, ttl % 60]
     lapsed = "%d:%02d" % [(duration / 60).to_i, (duration % 60).to_i]
+    saved = "%2.02f MB" % [(space_in - space_out).to_f / (1024 * 1024).to_f]
+    remain = "%6d" % [startNodes.length - count]
 
-    puts "#{lapsed} | #{togo} | #{startNodes.length - count}"
+    puts "#{lapsed} | #{togo} | #{remain} | #{saved}"
+    puts "-----\n#{ARGV[0]}" if count % 150 == 0
   end
 
   post, id = x.split('.')
@@ -50,6 +55,7 @@ startNodes.each { | x |
 
     File.open(page, 'r') { | content |
       raw = content.read
+      space_in += raw.length
       set = parsefile Nokogiri::HTML(raw, &:noblanks)
       set[:reblog].each { | row |
         from, who, post = row
@@ -61,11 +67,16 @@ startNodes.each { | x |
     todel << page
   }
 
-  File.open(output, 'w') { | f |
-    f << [reblog, like].to_json
-  }
+  # Only collapse if there is data
+  if like.length > 0
+    payload = [reblog, like].to_json
+    space_out += payload.length
+    File.open(output, 'w') { | f |
+      f << payload
+    }
 
-  ## Make sure that we don't remove the file until
-  # after our intended output has been written
-  todel.each { | file | File.unlink(file) }
+    ## Make sure that we don't remove the file until
+    # after our intended output has been written
+    todel.each { | file | File.unlink(file) }
+  end
 }
